The MTU is handled as a 32-bit entity within the SCTP stack. This was reported by Peter Kasting from Google.
MFC after: 3 days


Former-commit-id: d53bb9d93677e7e77114349fce0537111df323af

+++ sys/netinet/sctp_os_bsd.h
@@ -1457,7 +1457,7 @@ name|sa
 parameter_list|,
 name|rt
 parameter_list|)
-value|((rt != NULL) ? rt->rt_mtu : 0)
+value|((uint32_t)((rt != NULL) ? rt->rt_mtu : 0))
 end_define
 
 begin_define
